+++ app/src/main/java/com/jdevs/timeo/data/Record.kt
@@ -8,7 +8,7 @@ import java.util.Date
 @Keep
 data class Record(
     val name: String,
-    val time: Int = 0,
+    val time: Long = 0,
     val activityId: String = "",
     @ServerTimestamp var timestamp: Date = Calendar.getInstance().time
 ) {
